+++ src/main/java/ballistix/client/screen/ScreenSAMTurret.java
@@ -71,10 +71,10 @@ public class ScreenSAMTurret extends GenericScreen<ContainerSAMTurret> {
             width = font.width(radar);
 
             if(width > 100) {
-                scale = 100.0F / (float) width;
+                scale = 100.0F / width;
             }
 
-            float remHeight = (height - (float) height * scale) / 2.0F;
+            float remHeight = (height - height * scale) / 2.0F;
 
             graphics.pose().pushPose();
