+++ src/main/java/ballistix/prefab/screen/WrapperESMTowerDetections.java
@@ -75,8 +75,8 @@ public class WrapperESMTowerDetections {
                 int moveRoom = screen.slider.height - 15 -2;
 
                 // int moveRoom = 102 - 2;
-                double moved = (double) topRowIndex / (double) (lastRowCount - (double) BUTTON_COUNT);
-                slider.setSliderYOffset((int) ((double) moveRoom * moved));
+                double moved = topRowIndex / (lastRowCount - (double) BUTTON_COUNT);
+                slider.setSliderYOffset((int) (moveRoom * moved));
             }
         } else {
             slider.updateActive(false);
@@ -120,8 +120,8 @@ public class WrapperESMTowerDetections {
                     double heightRatio = (double) mouseHeight / (double) sliderHeight;
                     topRowIndex = (int) Math.round((lastRowCount - BUTTON_COUNT) * heightRatio);
                     int moveRoom = slider.height - 15 - 2;
-                    double moved = (double) topRowIndex / (double) (lastRowCount - (double)BUTTON_COUNT);
-                    slider.setSliderYOffset((int) ((double) moveRoom * moved));
+                    double moved = topRowIndex / (lastRowCount - (double)BUTTON_COUNT);
+                    slider.setSliderYOffset((int) (moveRoom * moved));
                 }
             }
         };
